+++ inc/rql/sock.h
@@ -48,7 +48,7 @@ union rql_sock_u
 
 struct rql_sock_s
 {
-    uint64_t ref;
+    uint32_t ref;
     uint32_t n;     /* buffer n */
     uint32_t sz;    /* buffer sz */
     rql_sock_e tp;
